인증, 보안 Api형태로 진행하도록 수정

+++ src/main/java/com/example/auth/domain/post/post/controller/ApiV1PostController.java
@@ -6,13 +6,17 @@ import com.example.auth.domain.post.post.dto.PostDto;
 import com.example.auth.domain.post.post.entity.Post;
 import com.example.auth.domain.post.post.service.PostService;
 import com.example.auth.global.dto.RsData;
+import com.example.auth.global.exception.ServiceException;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/posts")
@@ -21,6 +25,7 @@ public class ApiV1PostController {
 
     private final PostService postService;
     private final MemberService memberService;
+    private final HttpServletRequest request;
 
     @GetMapping
     public RsData<List<PostDto>> getItems() {
@@ -50,25 +55,44 @@ public class ApiV1PostController {
         );
     }
 
+    record DeleteReqBody(@NotNull Long authorId,
+                         @NotBlank @Length(min = 3) String password
+    ) {
+    }
+
     @DeleteMapping("/{id}")
     public RsData<Void> delete(@PathVariable long id) {
+
+        Member actor = getAuthenticatedActor();
         Post post = postService.getItem(id).get();
+
+        if (post.getAuthor().getId() != actor.getId()) {
+            throw new ServiceException("403-1", "자신이 작성한 글만 삭제 가능합니다.");
+        }
+
         postService.delete(post);
 
         return new RsData<>(
-                "204-1",
+                "200-1",
                 "%d번 글 삭제가 완료되었습니다.".formatted(id)
         );
     }
 
 
-    record ModifyReqBody(@NotBlank @Length(min = 3) String title, @NotBlank @Length(min = 3) String content) {
+    record ModifyReqBody(@NotBlank @Length(min = 3) String title,
+                         @NotBlank @Length(min = 3) String content) {
     }
 
     @PutMapping("{id}")
     public RsData<Void> modify(@PathVariable long id, @RequestBody @Valid ModifyReqBody body) {
 
+        Member actor = getAuthenticatedActor();
         Post post = postService.getItem(id).get();
+
+        if (post.getAuthor().getId() != actor.getId()) {
+            throw new ServiceException("403-1", "자신이 작성한 글만 수정 가능합니다.");
+        }
+
         postService.modify(post, body.title(), body.content());
         return new RsData<>(
                 "200-1",
@@ -78,25 +102,33 @@ public class ApiV1PostController {
     }
 
 
-    record WriteReqBody(@NotBlank @Length(min = 3) String title, @NotBlank @Length(min = 3) String content) {
-    }
-
-    record WriteResBody(long id, long totalCount) {
+    record WriteReqBody(@NotBlank @Length(min = 3) String title,
+                        @NotBlank @Length(min = 3) String content) {
     }
 
     @PostMapping
-    public RsData<WriteResBody> write(@RequestBody @Valid WriteReqBody body) {
+    public RsData<PostDto> write(@RequestBody @Valid WriteReqBody body) {
 
-        Member actor = memberService.findByUsername("user3").get();
+        Member actor = getAuthenticatedActor();
         Post post = postService.write(actor, body.title(), body.content());
 
         return new RsData<>(
                 "200-1",
                 "글 작성이 완료되었습니다.",
-                new WriteResBody(
-                        post.getId(),
-                        postService.count()
-                )
+                new PostDto(post)
         );
+    }
+
+    private Member getAuthenticatedActor() {
+
+        String authorizationValue = request.getHeader("Authorization");
+        String apiKey = authorizationValue.substring("Bearer ".length());
+        Optional<Member> opActor = memberService.findByApiKey(apiKey);
+
+        if(opActor.isEmpty()) {
+            throw new ServiceException("401-1", "잘못된 비밀번호 입니다.");
+        }
+
+        return opActor.get();
     }
 }
